Extract the game to its own package
